Money transfer

+++ accounts/src/main/java/lt/rieske/accounts/domain/MoneyDepositedEvent.java
@@ -6,8 +6,8 @@ import lt.rieske.accounts.eventsourcing.Event;
 
 @Value
 public class MoneyDepositedEvent implements Event<Account> {
-    private final int amountDeposited;
-    private final int balance;
+    private final long amountDeposited;
+    private final long balance;
 
     @Override
     public void apply(Account aggregate) {
